Day 16 - Again, now covering more edge cases
The previous version worked but relied on the endpoint being in the top
right corner (and therefore the two approaches to it having different
scores). If the maze was designed so that two equal routes could meet at
the endpoint, only one would be counted. eg

    #######
    #  E  #
    # ### #
    #     #
    ### ###
    #S  ###
    #######

Should be 4010 and 16 for parts one and two respectively.